Add simple number of kills and winPlace label

+++ app/src/main/kotlin/es/voghdev/playbattlegrounds/features/matches/api/model/ParticipantStatsApiEntry.kt
@@ -35,7 +35,7 @@ class ParticipantStatsApiEntry(
         val name: String?,
         val playerId: String?,
         val revives: Int?,
-        val rideDistance: Int?,
+        val rideDistance: Float?,
         val roadKills: Int?,
         val teamKills: Int?,
         val timeSurvived: Int?,
